+++ scilab/modules/fileio/src/cpp/mget.cpp
@@ -104,7 +104,7 @@ void C2F(mgetnc)(int* fd, void* res, int* n1, char* type, int* ierr)
             MGET_GEN_NC(int, c1);
             break;
         case 'l' :
-            MGET_GEN_NC(int32_t, c1);
+            MGET_GEN_NC(long long, c1);
             break;
         case 's' :
             MGET_GEN_NC(short, c1);
@@ -125,7 +125,7 @@ void C2F(mgetnc)(int* fd, void* res, int* n1, char* type, int* ierr)
                     MGET_GEN_NC(unsigned int, c2);
                     break;
                 case 'l' :
-                    MGET_GEN_NC(uint32_t, c2);
+                    MGET_GEN_NC(unsigned long long, c2);
                     break;
                 case 's' :
                     MGET_GEN_NC(unsigned short, c2);
@@ -184,7 +184,7 @@ void mget2(FILE *fa, int swap, double *res, int n, char *type, int *ierr)
             MGET_GEN(int, c1);
             break;
         case 'l' :
-            MGET_GEN(int32_t, c1);
+            MGET_GEN(long long, c1);
             break;
         case 's' :
             MGET_GEN(short, c1);
@@ -205,7 +205,7 @@ void mget2(FILE *fa, int swap, double *res, int n, char *type, int *ierr)
                     MGET_GEN(unsigned int, c2);
                     break;
                 case 'l' :
-                    MGET_GEN(uint32_t, c2);
+                    MGET_GEN(unsigned long long, c2);
                     break;
                 case 's' :
                     MGET_GEN(unsigned short, c2);
